SED: Adding a option to select which physical component (plot series) be enabled or disabled.
Use the --series <series1 series2...> option to indicate one o more of this series:
stellar_attenuated,
stellar_unattenuated,
nebular,
dust,
agn,
radio,
model.

If the --seriesdisabled flag is used, then all available series will be enabled, except the ones that are indicated in the --series option. If not, then only the series indicated in the --series option will be enabled.